\name{FITS file creation}
\alias{createFITSFile}
\title{Create a FITS file}
\usage{createFITSFile(file.name)}
\arguments{
	\item{file.name}{A string containing the name of the file to create.}
}
\author{Maurizio Tomasi <maurizio.tomasi@unimi.it>}
\details{
}
\value{
A object of type \code{cfitsio_obj}.
}
